+++ sdk/boletoPayment/boletoPayment.js
@@ -13,7 +13,7 @@ class BoletoPayment extends Resource {
      * to the Stark Bank API and returns the list of created objects.
      *
      * Parameters (conditionally required):
-     * @param line [string, default null]: Number sequence that describes the payment. Either 'line' or 'bar_code' parameters are required. If both are sent, they must match. ex: '34191.09008 63571.277308 71444.640008 5 81960000000062'
+     * @param line [string, default null]: Number sequence that describes the payment. Either 'line' or 'barCode' parameters are required. If both are sent, they must match. ex: '34191.09008 63571.277308 71444.640008 5 81960000000062'
      * @param barCode [string, default null]: Bar code number that describes the payment. Either 'line' or 'barCode' parameters are required. If both are sent, they must match. ex: '34195819600000000621090063571277307144464000'
      *
      * Parameters (required):
@@ -21,13 +21,13 @@ class BoletoPayment extends Resource {
      * @param description [string]: Text to be displayed in your statement (min. 10 characters). ex: 'payment ABC'
      *
      * Parameters (optional):
+     * @param amount [int, default null]: amount to be paid. If none is informed, the current boleto value will be used. ex: 23456 (= R$ 234.56)
      * @param scheduled [string, default today]: payment scheduled date. ex: '2020-03-10'
      * @param tags [list of strings]: list of strings for tagging
      *
      * Attributes (return-only):
      * @param id [string, default null]: unique id returned when payment is created. ex: '5656565656565656'
      * @param status [string, default null]: current payment status. ex: 'success' or 'failed'
-     * @param amount [int, default null]: amount automatically calculated from line or bar_code. ex: 23456 (= R$ 234.56)
      * @param fee [integer, default null]: fee charged when boleto payment is created. ex: 200 (= R$ 2.00)
      * @param created [string, default null]: creation datetime for the payment. ex: '2020-03-10 10:30:00.000'
      *