+++ sdk/workspace/workspace.js
@@ -13,7 +13,9 @@ class Workspace extends Resource {
      * Parameters (required):
      * @param username [string]: Simplified name to define the workspace URL. This name must be unique across all Stark Bank Workspaces. Ex: 'starkbankworkspace'
      * @param name [string]: Full name that identifies the Workspace. This name will appear when people access the Workspace on our platform, for example. Ex: 'Stark Bank Workspace'
-     * @param allowedTaxIds [list of strings]: list of tax IDs that will be allowed to send Deposits to this Workspace. ex: ['012.345.678-90', '20.018.183/0001-80']
+     * 
+     * Parameters (optional):
+     * @param allowedTaxIds [list of strings, default null]: list of tax IDs that will be allowed to send Deposits to this Workspace. If empty, all are allowed. ex: ['012.345.678-90', '20.018.183/0001-80']
      * 
      * Attributes:
      * @param id [string, default null]: unique id returned when the workspace is created. ex: '5656565656565656'